cleaned up export view code
