+++ escalate/core/views/export_views.py
@@ -4,9 +4,11 @@ from django.contrib.auth.mixins import LoginRequiredMixin
 
 import core.models
 import core.forms
-from core.views.crud_view_methods.model_view_generic import GenericModelExport
-from core.views.exports import export_methods
-from core.views.exports import file_types as export_file_types
+from core.views.exports import (
+    GenericModelExport,
+    export_methods,
+    file_types as export_file_types
+)
 
 class LoginRequired(LoginRequiredMixin):
     login_url = '/'
@@ -14,9 +16,9 @@ class LoginRequired(LoginRequiredMixin):
 
 
 def create_export_view(model_name, methods):
-    for t in export_file_types.file_types:
-        methods['file_type'] = t
-        class_name = f'{model_name}Export{t.capitalize()}'
+    for file_type in export_file_types:
+        methods['file_type'] = file_type
+        class_name = f'{model_name}Export{file_type.capitalize()}'
         globals()[class_name] = type(class_name,
                                           tuple([LoginRequired, GenericModelExport]), methods)
